최종 커밋

+++ src/main/java/com/hanghae/my_blog/service/CommentService.java
@@ -6,10 +6,13 @@ import com.hanghae.my_blog.dto.CompleteResponseDto;
 import com.hanghae.my_blog.entity.Comment;
 import com.hanghae.my_blog.entity.Post;
 import com.hanghae.my_blog.entity.User;
+import com.hanghae.my_blog.entity.UserRoleEnum;
 import com.hanghae.my_blog.jwt.JwtUtil;
+import com.hanghae.my_blog.repository.CommentLikesRepository;
 import com.hanghae.my_blog.repository.CommentRepository;
 import com.hanghae.my_blog.repository.PostRepository;
 import com.hanghae.my_blog.repository.UserRepository;
+import com.hanghae.my_blog.util.UserUtil;
 import io.jsonwebtoken.Claims;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -25,69 +28,65 @@ public class CommentService {
     private final PostRepository postRepository;
     private final CommentRepository commentRepository;
     private final UserRepository userRepository;
+    private final UserUtil userUtil;
+    private final CommentLikesRepository commentLikesRepository;
 
+    @Transactional
     //댓글 저장하기
-    public CommentResponseDto saveComment(Long postid, CommentRequestDto commentRequestDto, HttpServletRequest request) {
+    public CommentResponseDto saveComment(Long postId, CommentRequestDto commentRequestDto, HttpServletRequest httpServletRequest) {
         //로그인 여부 확인
-        User user = tokenChecking(request);
+        User user = userUtil.getUserInfo(httpServletRequest);
         //게시글 저장 여부 확인
-        Post post = postRepository.findById(postid).orElseThrow(
-                () -> new NullPointerException("게시글이 존재하지 않습니다")
+        Post post = postRepository.findById(postId).orElseThrow(
+                () -> new IllegalArgumentException("게시글이 존재하지 않습니다")
         );
         //댓글 저장
         Comment comment = new Comment(commentRequestDto,  post, user);
         commentRepository.save(comment);
-        return new CommentResponseDto(comment.getContent());
+        return new CommentResponseDto(comment);
     }
     @Transactional
-    public CommentResponseDto updateComment(Long postid, Long commentid, CommentRequestDto commentRequestDto, HttpServletRequest request) {
+    public CommentResponseDto updateComment(Long postId, Long commentId, CommentRequestDto commentRequestDto, HttpServletRequest httpServletRequest) {
         //로그인 여부 확인
-        User user = tokenChecking(request);
+        User user = userUtil.getUserInfo(httpServletRequest);
         //게시글 존재 여부 확인
-        Post post = postRepository.findById(postid).orElseThrow(
-                () -> new NullPointerException(("게시글이 존재하지 않습니다")
-        ));
+        if(!postRepository.existsById(postId)){
+            throw new IllegalArgumentException("게시글이 존재하지 않습니다");
+        }
         //댓글 존재 여부 확인
-        Comment comment = commentRepository.findById(commentid).orElseThrow(
-                () -> new NullPointerException(("댓글이 존재하지 않습니다")
+        Comment comment = commentRepository.findById(commentId).orElseThrow(
+                () -> new IllegalArgumentException(("댓글이 존재하지 않습니다")
         ));
-        //댓글 업데이트
+        //ADMIN 권한, username 확인 후 댓글 업데이트
+        if(comment.getUser().getUsername().equals(user.getUsername()) || user.getRole() == UserRoleEnum.ADMIN){
         comment.update(commentRequestDto);
+        }else{
+            throw new IllegalArgumentException("올바른 사용자가 아닙니다");
+        }
+        Long commentCnt = commentLikesRepository.countByCommentAndLikeCheckIsTrue(comment);
         //수정된 댓글 반환
-        return new CommentResponseDto(comment.getContent());
+        return new CommentResponseDto(comment, commentCnt);
     }
+    @Transactional
     //댓글 삭제하기
-    public CompleteResponseDto deleteComment(Long postid, Long commentid, HttpServletRequest request) {
+    public CompleteResponseDto deleteComment(Long postId, Long commentId, HttpServletRequest httpServletRequest) {
         //로그인 여부 확인
-        User user = tokenChecking(request);
+        User user = userUtil.getUserInfo(httpServletRequest);
         //게시글 저장 여부 확인
-        Post post = postRepository.findById(postid).orElseThrow(
-                () -> new NullPointerException(("게시글이 존재하지 않습니다")
-        ));
+        if(!postRepository.existsById(postId)){
+            throw new IllegalArgumentException("게시글이 존재하지 않습니다");
+        }
         //댓글 저장 여부 확인
-        Comment comment = commentRepository.findById(commentid).orElseThrow(
-                () -> new NullPointerException(("댓글이 존재하지 않습니다")
+        Comment comment = commentRepository.findById(commentId).orElseThrow(
+                () -> new IllegalArgumentException(("댓글이 존재하지 않습니다")
         ));
-        //댓글 삭제
-        commentRepository.delete(comment);
+        //ADMIN 권한, username 확인 후 댓글 삭제
+        if(comment.getUser().getUsername().equals(user.getUsername()) || user.getRole() == UserRoleEnum.ADMIN){
+            commentRepository.delete(comment);
+        }else{
+            throw new IllegalArgumentException("올바른 사용자가 아닙니다");
+        }
         //삭제 완료 반환
         return new CompleteResponseDto("삭제 완료");
-    }
-    //로그인 여부 확인
-    public User tokenChecking(HttpServletRequest request) {
-        String token = jwtUtil.resolveToken(request);
-        Claims claims;
-        if (token != null) {
-            if (jwtUtil.validateToken(token)) {
-                claims = jwtUtil.getUserInfoFromToken(token);
-            } else {
-                throw new IllegalArgumentException("Token Error");
-            }
-            User user = userRepository.findByUsername(claims.getSubject()).orElseThrow(
-                    () -> new NullPointerException(("사용자가 존재하지 않습니다")
-            ));
-            return user;
-        }throw new IllegalArgumentException("로그인이 필요합니다");
-
     }
 }
